slurmctld/statistics - Widen per-job counters in job_stats_t
cpus_alloc is copied from job_record_t.total_cpus and min_nodes and
max_nodes from job_details_t, all of which are uint32_t. Storing them
as uint16_t truncated any job asking for more than 65535 CPUs or nodes,
which then propagated into the aggregated job and partition metrics.

+++ src/slurmctld/statistics.h
@@ -189,14 +189,14 @@ typedef struct scheduling_statistics {
 
 typedef struct job_statistics {
 	char *account;
-	uint16_t cpus_alloc;
+	uint32_t cpus_alloc;
 	uint64_t gpus_alloc;
 	uint32_t job_array_cnt; /* If job array and PD, number of array tasks */
 	uint32_t job_id;
 	uint32_t job_state;
-	uint16_t max_nodes;
+	uint32_t max_nodes;
 	uint64_t memory_alloc;
-	uint16_t min_nodes;
+	uint32_t min_nodes;
 	bitstr_t *node_bitmap;
 	char *partition;
 	uint32_t state_reason;
